Restructure project: move all days under src/main/java, add resources folder, remove old modules
