+++ .gitignore
@@ -1,3 +1,6 @@
+ ./idea
+ *.iml
+
 # Compiled class file
 *.class
