--- aoc2022/.idea/encodings.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="Encoding">
-    <file url="file://$PROJECT_DIR$/day1/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/day1/src/main/resources" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/day2/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/day2/src/main/resources" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/day3/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/day3/src/main/resources" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/day4/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/day4/src/main/resources" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/src/main/resources" charset="UTF-8" />
-  </component>
-</project>
