--- aoc2022/day1/src/main/java/com/ivona/aoc/Elf.java
@@ -1,9 +1,0 @@
-package com.ivona.aoc;
-
-public record Elf(int position, int numberOfCalories) {
-    @Override
-    public String toString() {
-        return
-                "Elf #" + position + " with " + numberOfCalories + " calories ";
-    }
-}
