--- aoc2022/day1/src/main/resources/example.txt
@@ -1,14 +1,0 @@
-1000
-2000
-3000
-
-4000
-
-5000
-6000
-
-7000
-8000
-9000
-
-10000
