--- aoc2022/day2/src/main/resources/example.txt
@@ -1,3 +1,0 @@
-A Y
-B X
-C Z
