--- aoc2022/day4/src/main/resources/example.txt
@@ -1,6 +1,0 @@
-2-4,6-8
-2-3,4-5
-5-7,7-9
-2-8,3-7
-6-6,4-6
-2-6,4-8
