+++ src/main/resources/day2.txt
