+++ src/main/resources/day3.txt
