+++ src/main/resources/day4.txt
